+++ Item.py
@@ -1,12 +1,11 @@
 #Nick Mills
 # Created: 2015-05-18
-# Updated: 2015-05-21
-
-import random
+# Updated: 2019-03-24
 
 from Player import Player
 from Chance import chance
 from Enemy import Enemy
+from Utils import Utils
 
 class Item():
     """Base class for all Items"""
@@ -15,113 +14,73 @@ class Item():
         self.description = description
         self.value = value
 
-    def __str__(self):
-        return "\n{}\n=====\n{}\nValue: {}".format(self.name, self.description, self.value)
+    @staticmethod
+    def create_from_filedata() -> list:
+        """
+        ---------------------------------------------
+        Creates a list of Items from the cached filedata 
+        that was previously read.
+        Use: items = Item.create_from_filedata()
+        ---------------------------------------------
+        Parameters:
+            None
+        Returns:
+            items - list of Items created from cached 
+                      filedata (list of Item)
+        ---------------------------------------------
+        """
+        data = Utils.get_data()
+        item_data_start = 0
+        index = 0
+        while data[index] is not None:
+            index += 1
+        item_data_start = index + 1
+        items = []
+        item_data = data[item_data_start:]
+        dicts = item_data[0][0]
+        list_weapons = dicts['Weapons']
+        list_s_weapons = dicts['Super Weapons']
+        list_healers = dicts['Healers']
+        list_key_items = dicts['Key Items']
+        for weapon in list_weapons:
+            name = weapon['name']
+            desc = weapon['description']
+            val = weapon['value']
+            dmg = weapon['damage']
+            weapon = Weapon(name, desc, val, dmg)
+            items.append(weapon)
+        for s_weapon in list_s_weapons:
+            name = s_weapon['name']
+            desc = s_weapon['description']
+            val = s_weapon['value']
+            dmg = s_weapon['damage']
+            b_dmg = s_weapon['bonus_damage']
+            superweapon = SuperWeapon(name, desc, val, \
+                dmg, b_dmg)
+            items.append(superweapon)
+        for healer in list_healers:
+            name = healer['name']
+            desc = healer['description']
+            val = healer['value']
+            healing = healer['healing']
+            life_increase = healer['life_increase']
+            heal = Healer(name, desc, val, healing, life_increase)
+            items.append(heal)
+        for key in list_key_items:
+            name = key['name']
+            desc = key['description']
+            val = key['value']
+            key_item = KeyItem(name, desc, val)
+            items.append(key_item)
+        return items
 
-class KeyItems(Item):
+class KeyItem(Item):
     """Base class for all Key Items"""
     def __init__(self, name, description, value):
         super().__init__(name, description, value)
     
     def __str__(self):
-        return "\n{}\n*****\n{}\nValue: {}".format(self.name, self.description, self.value)
-    
-class CellarKey(KeyItems):
-    def __init__(self):
-        super().__init__(name="Cellar Key",
-                         description="A small key to unlock a cellar.",
-                         value=0)
-
-class Torch(KeyItems):
-    def __init__(self):
-        super().__init__(name="Torch",
-                         description="A torch that lets you see in the dark. Somehow does not burn out.",
-                         value=0)
-        
-class Book(KeyItems):
-    def __init__(self):
-        super().__init__(name="Book",
-                         description="A book name \"The History of the Skeleton King\". Of no use to you.",
-                         value=0)
-        
-class GreatOakKey(KeyItems):
-    def __init__(self):
-        super().__init__(name="Great Oak Key",
-                         description="A large diamond key that has a picture of the Great Oak on it.",
-                         value=0)
-        
-class Note(KeyItems):
-    def __init__(self):
-        super().__init__(name="Note",
-                         description="A note from your father. He hints about a Great Tree you where to find something that someone is missing. \"It is vitally important\" He explains.",
-                         value=0)
-        
-class JewelEgg(KeyItems):
-    def __init__(self):
-        super().__init__(name="Jewel Egg",
-                         description="A jewel-encrested egg. Surprisingly heavy for how small it is.",
-                         value=0)
-    
-class LegendarySword(KeyItems):
-    def __init__(self):
-        super().__init__(name="Legendary Sword",
-                         description="The sword of legend. Legend has it that when put through something into a hidden key hole, the passage to Shangri-La will be opened",
-                         value=0)
-        
-class TriForce(KeyItems):
-    def __init__(self):
-        super().__init__(name="TriForce",
-                         description="A triangle seperated by smaller triangles. The triangle in the middle is missing however, so there is a hole in it.",
-                         value=0)
-        
-class Healer(Item):
-    """Base class for Potions"""
-    def __init__(self, name, description, value, healing, life_increase):
-        self.healing = healing
-        self.life_increase = life_increase
-        super().__init__(name, description, value)
-
-    def ___str___(self):
-        return "\n{}\n=====\n{}\nValue: {}\nHealth Gained: {}\nLife Increase: {}".format(self.name, self.description, self.value, self.healing, self.life_increase)
-
-    def heal(player, healing):
-        player.hp += healing
-        return player.hp
-
-    def increase_life(player, life_increase):
-        player.hp = player.max_hp
-        player.max_hp += life_increase
-        player.hp = player.max_hp
-        return player.hp, player.max_hp
-    
-    def full_restore(player):
-        player.hp = player.max_hp
-        return player.hp, player.max_hp
-    
-class Potion(Healer):
-    def __init__(self):
-        super().__init__(name="Potion",
-                         description="A red potion that heals 50 hp. Single-Use only.",
-                         value=15,
-                         healing=50,
-                         life_increase=0)
-
-class FullPotion(Healer):
-        def __init__(self):
-            super().__init__(name="Full Potion",
-                             description="A orange potion that heals you to max hp! Single-Use only.",
-                             value=40,
-                             healing=0, #0 is used as a placeholder, the healing is actually Player.max_hp but the full_restore function in Healer handles that
-                             life_increase=0)
-
-class LifePotion(Healer):
-    def __init__(self):
-        super().__init__(name="Life Potion",
-                         description="A potion with a gold color. Increases HP by 100.",
-                         value=50,
-                         healing=0,
-                         life_increase=100)
-
+        return "\n{}\n".format(self.name)+("*"*len(self.name))+"\n{}\nValue: {}".format(self.description, self.value)
 
 class Weapon(Item):
     """Base class for all Weapons"""
@@ -130,110 +89,33 @@ class Weapon(Item):
         super().__init__(name, description, value)
 
     def __str__(self):
-        return "\n{}\n=====\n{}\nValue: {}\nDamage: {}\n".format(self.name, self.description, self.value, self.damage)
+        return "\n{}\n".format(self.name)+("="*len(self.name))+"\n{}\nValue: {}\nDamage: {}\n".format(self.description, self.value, self.damage)
 
 class SuperWeapon(Weapon):
     """Base class for all Legendary Weapons"""
-    def __init__(self, name, description, value, damage, magic_dmg, poison_dmg, bleed_dmg):
-        self.magic_dmg = magic_dmg
-        self.poison_dmg = poison_dmg
-        self.bleed_dmg = bleed_dmg
-        super().__init__(name, description, value, damage)
+    def __init__(self, name, description, value, damage, bonus_damage):
+        super().__init__(name, description, value, damage + bonus_damage)
 
-class Rock(Weapon):
-    def __init__(self):
-        super().__init__(name="Rock",
-                         description="A sizeable rock that can easily smash someone's skull in.",
-                         value=0,
-                         damage=5)
+class Healer(Item):
+    """Base class for Potions"""
+    def __init__(self, name, description, value, healing, life_increase):
+        self.healing = healing
+        self.life_increase = life_increase
+        super().__init__(name, description, value)
 
-class Dagger(Weapon):
-    def __init__(self):
-        super().__init__(name="Dagger",
-                       description="A shiny dagger with some rust. Somewhat more dangerous than a rock.",
-                       value=10,
-                       damage=10)
+    def __str__(self):
+        return "\n{}\n".format(self.name)+("="*len(self.name))+"\n{}\nValue: {}\nHealth Gained: {}\nLife Increase: {}".format(self.description, self.value, self.healing, self.life_increase)
 
-class Sword(Weapon):
-    def __init__(self):
-        super().__init__(name="Sword",
-                         description="An average sized sword. Deadly in the hands of a swordsman.",
-                         value=15,
-                         damage=20)
+    def heal(self, player, healing):
+        player.hp += healing
+        return player.hp
 
-class HeavySword(Weapon):
-    def __init__(self):
-        super().__init__(name="Heavy Sword",
-                         description="A large sword that requires two hands to swing",
-                         value=20,
-                         damage=25)
-
-class Bow(Weapon):
-    def __init__(self):
-        super().__init__(name="Bow",
-                         description="A huntsman's bow. Deadly at a range, useless at point blank.",
-                         value=15,
-                         damage=25)
-
-class Crossbow(Weapon):
-    def __init__(self):
-        super().__init__(name="Crossbow",
-                         description="Two handed crossbow. Deadly at all ranges. More effective at a larger range than a huntsman's bow.",
-                         value=20,
-                         damage=30)
-
-class Handcannon(Weapon):
-    def __init__(self):
-        super().__init__(name="Handcannon",
-                         description="A cannon that fits in your hand. Deadly. End. Of. Story.",
-                         value=50,
-                         damage=50)
-
-class Krambit(Weapon):
-    def __init__(self):
-        super().__init__(name="Krambit",
-                         description="Small knife with a curved edge. Worth $262 with no skins. Lots of skins available. \"Such skins, very wow\"",
-                         value=262,
-                         damage=70)
-
-class SwordOfDragox(SuperWeapon):
-    def __init__(self):
-        super().__init__(name="Sword Of Dragox",
-                       description="The sword of your father, Dragox. It hums with a mystical energy.",
-                       value=1000,
-                       damage=100,
-                       magic_dmg=50,
-                       poison_dmg=0,
-                       bleed_dmg=0)
-        
-    def total_damage(damage, magic_dmg):
-        damage = damage + magic_dmg
-        return damage
-
-class ScepterOfRomox(SuperWeapon):
-    def __init__(self):
-        super().__init__(name="Scepter of Romox",
-                         description="The scepter of your father's father, Romox. Drips with poison.",
-                         value=2500,
-                         damage=200,
-                         magic_dmg=0,
-                         poison_dmg=150,
-                         bleed_dmg=0)
-        
-        def total_damage(damage, poison_dmg):
-            damage = damage + poison_dmg
-            return damage
-
-class UmariOfFencor(SuperWeapon):
-    def __init__(self):
-        super().__init__(name="Umari of Fencor",
-                         description="The umari of your uncle, Fencor. It is a whip with small daggers at the ends that is covered in blood. Hard to master.",
-                         value=10000,
-                         damage=500,
-                         magic_dmg=0,
-                         poison_dmg=0,
-                         bleed_dmg=400)
-        
-        def total_damage(damage, bleed_dmg):
-            damage = damage + bleed_dmg
-            return damage
+    def increase_life(self, player, life_increase):
+        player.hp = player.max_hp
+        player.max_hp += life_increase
+        player.hp = player.max_hp
+        return player.hp, player.max_hp
+    
+    def full_restore(self, player):
+        player.hp = player.max_hp
+        return player.hp, player.max_hp